use short instead of int to present type of Factor to save memory.

+++ src/reg/factor/Factor.java
@@ -12,5 +12,5 @@ public interface Factor {
     /**
      * @return the type of the instance. type reference {@link FactorTypeRegister}
      */
-    int type();
+    short type();
 }
